Add files via upload
This Image Classifier was trained through a Neural Network using supervised learning. There are comments throughout the code in which describe what I am doing throughout but I am cleaning and training the data until it finds the most accurate model and saves it so we can use it in cnn--inference. The training data set is too large but can be found at: https://www.kaggle.com/datasets/puneet6060/intel-image-classification?resource=download